// orders/src/module/orders/interfaces/orders.interface.ts
export const OrderStatus = {
  Created: 'Created',
  Confirmed: 'Confirmed',
  Processing: 'Processing',
  Collected: 'Collected',
  Done: 'Done',
  Canceled: 'Canceled',
  UNRECOGNIZED: 'UNRECOGNIZED',
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];
// ...
  export type Created = typeof OrderStatus.Created;
  export type Confirmed = typeof OrderStatus.Confirmed;
  export type Processing = typeof OrderStatus.Processing;
  export type Collected = typeof OrderStatus.Collected;
  export type Done = typeof OrderStatus.Done;
  export type Canceled = typeof OrderStatus.Canceled;
  export type UNRECOGNIZED = typeof OrderStatus.UNRECOGNIZED;
// ...
export const SortDirection = { desc: 'desc', asc: 'asc', UNRECOGNIZED: 'UNRECOGNIZED' } as const;

export type SortDirection = (typeof SortDirection)[keyof typeof SortDirection];

export namespace SortDirection {
  export type desc = typeof SortDirection.desc;
  export type asc = typeof SortDirection.asc;
  export type UNRECOGNIZED = typeof SortDirection.UNRECOGNIZED;
}

export interface Empty {}

export interface SortQuery {
  field: string;
  direction: SortDirection;
}

export interface Item {
  productId: string;
  quantity: number;
  unitPrice: number;
  sum: number;
}

export interface OrderResponse {
  orderId: string;
  status: OrderStatus;
  date: number;
  totalSum: number;
  customerId: string;
  items: Item[];
  createdAt: number;
}

export interface FindAllOrderResponse {
  orderId: string;
  status: OrderStatus;
  date: number;
  totalSum: number;
  customerId: string;
  createdAt: number;
}

export interface FindAllOrdersResponse {
  pagesCount: number;
  page: number;
  pageSize: number;
  totalCount: number;
  items: FindAllOrderResponse[];
}

export interface CreateOrderRequest {
  customerId: string;
  items: Item[];
}

export interface OrderRequest {
  orderId: string;
  customerId: string;
}

export interface FindAllOrdersRequest {
  customerId: string;
  pageNumber?: number | undefined;
  pageSize?: number | undefined;
  sortBy: SortQuery[];
}